workspace: wasm32-unknown-unknown support — 20/36 -> 27/36 crates (op-okqo)
Goal: make wasm32-unknown-unknown a supported target for every workspace
member. Measured baseline before this commit was 20 PASS / 16 FAIL on
`cargo check -p <crate> --lib --target wasm32-unknown-unknown`; it is now 27/36.

CLASS E — a real portability bug, not a build-config issue (op-okqo.5, closed):
  outram-park-mpi's `COLL_CONTEXT_OFFSET: usize = 1 << 40` is a compile-time
  overflow wherever usize is 32-bit. Fixed by WIDENING the communicator context
  id to u64 throughout rather than truncating the constant: the value is a
  numeric namespace tag, not a memory size, and truncating would have silently
  collapsed the 40-bit separation the collective/point-to-point isolation
  argument depends on. outram-park-fork-pflotran needed no separate fix — it
  reached the same constant through mpi and went green with it.

  This would equally have failed on armv7-linux-androideabi and
  i686-linux-android, which the Android rule nominally covers but whose
  acceptance command names only aarch64. Filed against op-zfr.

CLASS A — rayon (op-okqo.1): rayon-core does not build for wasm at all. Five
crates affected: njoy-outram-park-fork, outram-mc-libs, boon-lay directly;
nee_soon and raffles transitively (both went green for free).

  Mechanism chosen deliberately: TARGET-gated off wasm only, NOT feature-gated
  and NOT gated off Android. rayon is pure Rust with no system component and is
  Android-in-scope, so feature-gating it would cost Termux its multi-core path
  for no reason. Same for ratatui/crossterm: no terminal in a browser, but a
  TUI is explicitly Android-in-scope, so it is gated off wasm alone.

  On wasm the call sites use a small in-crate `wasm_par` module whose adapters
  run the same work serially. This is not a claim of parallelism — the module
  docs say so plainly — and for these crates it is numerically exact: the Monte
  Carlo drivers seed per history and document thread-count-independent
  eigenvalues, and boon-lay's CPU Walk-on-Spheres path is already the trusted
  reference the GPU path is validated against.

Also gated off wasm: njoy's fs2 (flock has no meaning without a filesystem;
the lock/unlock calls are cfg'd to match, and the surrounding double-checked
acquire is unchanged because wasm32 is single-threaded and has no peer to race)
and ureq/zip (a browser cannot open raw sockets; `net-fetch` on wasm would need
the fetch API, and the manifest now says so rather than failing obscurely).

boon-lay's gpu module is now gated off wasm as well as Android — for a
type-system reason, not policy: wgpu's WebGPU backend holds Rc<Cell<u32>>, so
GpuContext is !Send there and the `static OnceLock<Option<GpuContext>>` cache
cannot compile. Reaching WebGPU from wasm needs a thread_local cache instead,
which is a real change rather than a gate.

Verification: 27/36 crates pass the wasm --lib check (was 20/36); every crate
touched still passes native --lib and aarch64-linux-android --lib with 0
errors; outram-park-mpi's 35 tests pass.

Remaining 9 and their classes are tracked under epic op-okqo:
  bedok, outram-blender          - faer default features (op-okqo.2)
  kovan, tampines, tuas, dt-engine - getrandom/uuid wasm backend (op-okqo.3)
  kovan-discovery/-metrics/-semantics - std::os::unix (op-okqo.4)

No token accounting on this clone.

+++ crates/outram-park-mpi/src/comm_mgmt.rs
@@ -107,7 +107,7 @@ impl Communicator {
         let mut ctx = None;
         for i in 0..n {
             if map[1 + 2 * i] as i32 == color {
-                ctx = Some(map[2 + 2 * i] as usize);
+                ctx = Some(map[2 + 2 * i] as u64);
                 break;
             }
         }
@@ -142,13 +142,13 @@ impl Communicator {
     /// Allocate a fresh context and broadcast it from rank 0 so the whole group
     /// agrees (used by [`dup`](Communicator::dup) and
     /// [`create_from_group`](Communicator::create_from_group)).
-    pub(crate) fn agree_new_context(&self) -> MpiResult<usize> {
+    pub(crate) fn agree_new_context(&self) -> MpiResult<u64> {
         let agreed = if self.rank() == 0 {
             let id = self.alloc_context() as i64;
             self.broadcast(Some(&[id]), 0)?
         } else {
             self.broadcast(None, 0)?
         };
-        Ok(agreed[0] as usize)
+        Ok(agreed[0] as u64)
     }
 }
